COMPMID-3443: Android R CTS 1.3 QASYMM8_SIGNED failure with MEAN on CpuAcc
- Properly perform the division of the sum of the elements by the interested dimension of the tensor

Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/3145
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>

+++ src/core/NEON/kernels/NEReductionOperationKernel.cpp
@@ -724,14 +724,13 @@ struct RedOpX_quantized
 
                 if(op == ReductionOperation::MEAN_SUM)
                 {
-                    res /= in_info.dimension(0);
+                    res /= static_cast<int32_t>(in_info.dimension(0));
                 }
                 else
                 {
                     // Subtract accumulated offsets
                     res -= (in_info.dimension(0) - 1) * iq_info.offset;
                 }
-
                 *reinterpret_cast<T *>(output.ptr()) = utils::cast::saturate_cast<T>(res);
             }
         }
